+++ js/meta-manager.js
@@ -413,4 +413,6 @@ class MetaManager {
 // Export for use in other scripts
 if (typeof module !== 'undefined' && module.exports) {
   module.exports = MetaManager;
+} else if (typeof window !== 'undefined') {
+  window.MetaManager = MetaManager;
 }